[Vim] Set 'hidden'
I went a while with this turned off. See no advantage, and in fact there
is a huge problem: each time a buffer is hidden and returns to view, the
file is read from disk again, meaning that the buffer's undo history is
lost. (No interest in turning persistent undo on.)