[Emacs] Improve structure and format of config
Added some more documentation/commentary.

Use org-indent-mode: In this minor mode, all lines are prefixed for
display with the necessary amount of space. Also headlines are prefixed
with additional stars, so that the amount of indentation shifts by two
spaces per level.

This mode affects ths display only, eliminating the need to indent
section content, which was a hassle.