[Fish] Track release branch of laravel/homestead
Unlike the master branch, the release branch always points to a release.
The documentation claims that the release branch always points to the
*latest* release, however, I have not found this to be the case. For
example, the latest release is currently 9.2.2, which is tagged on
master. The release branch, on the other hand, contains 9.2.1 as the
latest tag (plus a couple stray commits after the tagged commit).